Move aovcomponents into aov folder, order alphabetically
