+++ src/appleseed/renderer/kernel/aov/aovaccumulator.h
@@ -46,9 +46,9 @@
 // Forward declarations.
 namespace foundation    { class Image; }
 namespace foundation    { class Tile; }
+namespace renderer      { class AOVComponents; }
 namespace renderer      { class Frame; }
 namespace renderer      { class PixelContext; }
-namespace renderer      { class AOVComponents; }
 namespace renderer      { class ShadingComponents; }
 namespace renderer      { class ShadingPoint; }
 namespace renderer      { class ShadingResult; }
