+++ src/appleseed/renderer/kernel/aov/aovcomponents.h
