+++ src/appleseed/renderer/kernel/shading/shadingengine.cpp
@@ -32,7 +32,7 @@
 
 // appleseed.renderer headers.
 #include "renderer/kernel/aov/aovaccumulator.h"
-#include "renderer/kernel/shading/aovcomponents.h"
+#include "renderer/kernel/aov/aovcomponents.h"
 #include "renderer/kernel/shading/shadingcomponents.h"
 #include "renderer/kernel/shading/shadingcontext.h"
 #include "renderer/kernel/shading/shadingray.h"
