+++ src/appleseed/renderer/modeling/aov/albedoaov.cpp
@@ -31,7 +31,7 @@
 
 // appleseed.renderer headers.
 #include "renderer/kernel/aov/aovaccumulator.h"
-#include "renderer/kernel/shading/aovcomponents.h"
+#include "renderer/kernel/aov/aovcomponents.h"
 #include "renderer/kernel/shading/shadingcomponents.h"
 #include "renderer/kernel/shading/shadingresult.h"
 #include "renderer/modeling/aov/aov.h"
